UI inplementazioa, itemak, eraso sistema, animazio berriak, bizitza sistema, arazo konponketa
Bizitza eta inbentarioaren UI garapena.
Itemak sortzeko plantilla.
Sortutako itemak badaukate eragina jokalarian -> ezpata erasotzeko, arkua geziak botatzeko eta suargia(implementazioa falta da) bihotzak eta edabeak (bihotzUI aldatzen dute) geziak (UI garatzea falta da).
Suargiarentzat animazioak egin dira(bazegoen egituran maila berria)
Jokalariaren hil eta berpiztu egoerak inlpementatu animazioekin
kutxarekin grabitatea batzen zen arazoa konponduta

+++ Hieroglifikoak_Unity/Assets/Scripts/Etsaia.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Etsaia : MonoBehaviour {
 
-    public int bizitzaPuntuak;
+    public float bizitzaPuntuak;
     public GameObject hilPartikula;
 
 	// Use this for initialization
@@ -17,7 +17,7 @@ public class Etsaia : MonoBehaviour {
 		
 	}
 
-    public void KolpeaJaso(int minPuntuak)
+    public void KolpeaJaso(float minPuntuak)
     {
         bizitzaPuntuak -= minPuntuak;
         Debug.Log("Ouch. " + minPuntuak + "bizitzaPuntu gutxiago.");
